Introduced error and warning reporting functions, to normalise the format of errors and warnings. Definitions of functions go in duchamp.cc. Changed all code that reports errors/warnings to use these new functions.
Fixed a couple of bugs that affected the way 2D images were dealt with: 
ReconSearch now looks at how many dimensions there are in the data array
before choosing the dimension of the reconstruction, and the minChannels
test was improved for the case of minChannels=0.

Some minor additions made to the Guide as well.



git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@120 450e7745-72da-457f-8a74-b07dc0cda934